Refactor: move project to package-by-domain/feature structure, updated imports
