+++ backend/src/main/java/com/veggieshop/auth/AuthService.java
@@ -1,6 +1,6 @@
 package com.veggieshop.service;
 
-import com.veggieshop.dto.AuthDto;
+import com.veggieshop.auth.AuthDto;
 
 public interface AuthService {
     AuthDto.AuthResponse login(AuthDto.AuthRequest request);
