+++ backend/src/main/java/com/veggieshop/category/CategoryDto.java
