+++ backend/src/main/java/com/veggieshop/offer/Offer.java
