+++ backend/src/main/java/com/veggieshop/offer/OfferDto.java
