+++ backend/src/main/java/com/veggieshop/offer/OfferRepository.java
@@ -1,6 +1,6 @@
 package com.veggieshop.repository;
 
-import com.veggieshop.entity.Offer;
+import com.veggieshop.offer.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
