+++ backend/src/main/java/com/veggieshop/product/Product.java
